Change SDRAMFile to MemoryIO.
 - Renames `SDRAMFile` to `MemoryIO` and migrates it into
   `machine_controller.py`.
 - Adds a new `sdram_alloc_as_io` method to `MachineController` which
   allocs SDRAM and then returns a file-like view onto it.

Please read over @mossblaser.
