+++ rig/machine_control/tests/test_machine_controller.py
@@ -4,7 +4,8 @@ import struct
 from .test_scp_connection import SendReceive, mock_conn  # noqa
 
 from ..consts import DataType, SCPCommands, LEDAction
-from ..machine_controller import MachineController, SpiNNakerMemoryError
+from ..machine_controller import (
+    MachineController, SpiNNakerMemoryError, MemoryIO)
 from ..packets import SCPPacket
 
 
@@ -25,6 +26,12 @@ def controller_rw_mock():
 
     cn._read.side_effect = read_mock
 
+    return cn
+
+
+@pytest.fixture
+def mock_controller():
+    cn = mock.Mock(spec=MachineController)
     return cn
 
 
@@ -547,3 +554,109 @@ class TestMachineController(object):
 
         assert str((x, y)) in str(excinfo.value)
         assert str(size) in str(excinfo.value)
+
+    @pytest.mark.parametrize("x, y", [(0, 1), (3, 4)])
+    @pytest.mark.parametrize("app_id", [30, 33])
+    @pytest.mark.parametrize("size", [8, 200])
+    @pytest.mark.parametrize("tag", [0, 2])
+    @pytest.mark.parametrize("addr", [0x67000000, 0x61000000])
+    def test_sdram_alloc_and_open(self, app_id, size, tag, addr, x, y):
+        """Test allocing and getting a file-like object returned."""
+        # Create the mock controller
+        cn = MachineController("localhost")
+
+        cn._send_scp = mock.Mock()
+        cn._send_scp.return_value = SCPPacket(False, 1, 0, 0, 0, 0, 0, 0, 0, 0,
+                                              0x80, 0, addr, None, None, b"")
+
+        # Try the allocation
+        fp = cn.sdram_alloc_as_io(size, tag, x, y, app_id=app_id)
+
+        # Check the fp has the expected start and end
+        assert fp._start_address == addr
+        assert fp._end_address == addr + size
+
+        # Check the x and y are correct
+        assert fp._machine_controller is cn
+        assert fp._x == x
+        assert fp._y == y
+
+
+class TestMemoryIO(object):
+    """Test the SDRAM file-like object."""
+    @pytest.mark.parametrize("x, y", [(1, 3), (3, 0)])
+    @pytest.mark.parametrize("start_address", [0x60000000, 0x61000000])
+    @pytest.mark.parametrize("lengths", [[100, 200], [100], [300, 128, 32]])
+    def test_read(self, mock_controller, x, y, start_address, lengths):
+        sdram_file = MemoryIO(mock_controller, x, y,
+                              start_address, start_address+500)
+        assert sdram_file.tell() == 0
+
+        # Perform the reads, check that the address is progressed
+        calls = []
+        offset = 0
+        for n_bytes in lengths:
+            sdram_file.read(n_bytes)
+            assert sdram_file.tell() == offset + n_bytes
+            assert sdram_file.address == start_address + offset + n_bytes
+            calls.append(mock.call(x, y, 0, start_address + offset, n_bytes))
+            offset = offset + n_bytes
+
+        # Check the reads caused the appropriate calls to the machine
+        # controller.
+        mock_controller.read.assert_has_calls(calls)
+
+    def test_read_beyond(self, mock_controller):
+        sdram_file = MemoryIO(mock_controller, 0, 0,
+                              start_address=0, end_address=10)
+        sdram_file.read(100)
+        mock_controller.read.assert_called_with(0, 0, 0, 0, 10)
+
+        assert sdram_file.read(1) == b''
+        assert mock_controller.read.call_count == 1
+
+    @pytest.mark.parametrize("x, y", [(4, 2), (255, 1)])
+    @pytest.mark.parametrize("start_address", [0x60000004, 0x61000003])
+    @pytest.mark.parametrize("lengths", [[100, 200], [100], [300, 128, 32]])
+    def test_write(self, mock_controller, x, y, start_address, lengths):
+        sdram_file = MemoryIO(mock_controller, x, y,
+                              start_address, start_address+500)
+        assert sdram_file.tell() == 0
+
+        # Perform the reads, check that the address is progressed
+        calls = []
+        offset = 0
+        for i, n_bytes in enumerate(lengths):
+            n_written = sdram_file.write(chr(i % 256) * n_bytes)
+            assert n_written == n_bytes
+            assert sdram_file.tell() == offset + n_bytes
+            assert sdram_file.address == start_address + offset + n_bytes
+            calls.append(mock.call(x, y, 0, start_address + offset,
+                                   chr(i % 256) * n_bytes))
+            offset = offset + n_bytes
+
+        # Check the reads caused the appropriate calls to the machine
+        # controller.
+        mock_controller.write.assert_has_calls(calls)
+
+    def test_write_beyond(self, mock_controller):
+        sdram_file = MemoryIO(mock_controller, 0, 0,
+                              start_address=0, end_address=10)
+
+        assert sdram_file.write(b"\x00\x00" * 12) == 10
+
+        assert sdram_file.write(b"\x00") == 0
+        assert mock_controller.write.call_count == 1
+
+    @pytest.mark.parametrize("start_address", [0x60000004, 0x61000003])
+    @pytest.mark.parametrize("seeks", [(100, -3, 32, 5, -7)])
+    def test_seek(self, mock_controller, seeks, start_address):
+        sdram_file = MemoryIO(mock_controller, 0, 0,
+                              start_address, start_address+200)
+        assert sdram_file.tell() == 0
+
+        cseek = 0
+        for seek in seeks:
+            sdram_file.seek(seek)
+            assert sdram_file.tell() == cseek + seek
+            cseek += seek
